move abstract enums to their own file
